+++ frontend/src/constants/screenSize.ts
@@ -1,0 +1,6 @@
+const MOBILE = 400;
+const TABLET = 768;
+const DESKTOP = 1024;
+const DESKTOP_BIC = 1600;
+
+export default { MOBILE, TABLET, DESKTOP, DESKTOP_BIC };
